[Feat] 공통 UI RunningResultViewController로 정의
- 기존 RunningResultViewController는 SingleRunningResultViewController로 변경했습니다.
