+++ MateRunner/MateRunner/Presentation/HomeScene/ViewController/ResultViewController/RunningResultViewController.swift
@@ -2,21 +2,16 @@
 //  RunningResultViewController.swift
 //  MateRunner
 //
-//  Created by 김민지 on 2021/11/01.
+//  Created by 전여훈 on 2021/11/20.
 //
 
-import CoreLocation
-import MapKit
 import UIKit
 
-import RxCocoa
 import RxSwift
 
 class RunningResultViewController: UIViewController {
-    var viewModel: RunningResultViewModel?
-    private let disposeBag = DisposeBag()
     
-    lazy var contentView: UIView = {
+    var contentView: UIView = {
         let view = UIView()
         view.backgroundColor = .systemBackground
         return view
@@ -24,7 +19,7 @@ class RunningResultViewController: UIViewController {
     
     private lazy var scrollView = UIScrollView()
     
-    private lazy var closeButton: UIButton = {
+    lazy var closeButton: UIButton = {
         let button = UIButton()
         let xImage = UIImage(systemName: "xmark")
         button.setImage(xImage, for: .normal)
@@ -34,7 +29,7 @@ class RunningResultViewController: UIViewController {
         return button
     }()
     
-    private lazy var dateTimeLabel: UILabel = {
+    lazy var dateTimeLabel: UILabel = {
         let label = UILabel()
         label.text = "2020. 6. 10. - 오후 4:32"
         label.font = .notoSans(size: 18, family: .medium)
@@ -42,23 +37,25 @@ class RunningResultViewController: UIViewController {
         return label
     }()
     
-    private lazy var korDateTimeLabel: UILabel = {
+    lazy var korDateTimeLabel: UILabel = {
         let label = UILabel()
         label.text = "수요일 오후"
         label.font = .notoSans(size: 24, family: .medium)
         return label
     }()
     
-    private lazy var runningModeLabel: UILabel = {
+    lazy var runningModeLabel: UILabel = {
         let label = UILabel()
-        label.text = "혼자 달리기"
+        label.text = "hunihun956 메이트와 함께한 달리기 🏃🏻‍♂️"
         label.font = .notoSans(size: 24, family: .medium)
+        label.numberOfLines = 0
+        label.lineBreakMode = .byWordWrapping
         return label
     }()
     
     private lazy var upperSeparator = self.createSeparator()
     
-    private lazy var distanceLabel: UILabel = {
+    lazy var distanceLabel: UILabel = {
         let label = UILabel()
         label.text = "5.00"
         label.font = .notoSansBoldItalic(size: 64)
@@ -69,14 +66,14 @@ class RunningResultViewController: UIViewController {
         return label
     }()
     
-    private lazy var calorieLabel: UILabel = {
+    lazy var calorieLabel: UILabel = {
         let label = UILabel()
         label.text = "128"
         label.font = .notoSans(size: 24, family: .black)
         return label
     }()
     
-    private lazy var timeLabel: UILabel = {
+    lazy var timeLabel: UILabel = {
         let label = UILabel()
         label.text = "24:50"
         label.font = .notoSans(size: 24, family: .black)
@@ -89,15 +86,15 @@ class RunningResultViewController: UIViewController {
         timeLabel: self.timeLabel
     )
     
-    lazy var mapView = MKMapView()
-    
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.configureUI()
-        self.configureMap()
-        self.bindViewModel()
+        self.configureCommonUI()
     }
-    
+}
+
+// MARK: - Private Functions
+
+private extension RunningResultViewController {
     func createSeparator() -> UIView {
         let view = UIView()
         view.backgroundColor = .systemGray4
@@ -107,41 +104,8 @@ class RunningResultViewController: UIViewController {
         return view
     }
     
-    func configureDifferentSection() {
-        self.configureMapView()
-    }
-    
-    func configureMapView() {
-        self.contentView.addSubview(self.mapView)
-        
-        self.mapView.snp.makeConstraints { [weak self] make in
-            guard let self = self else { return }
-            
-            make.top.equalTo(self.myResultView.snp.bottom).offset(15)
-            make.left.equalToSuperview().offset(15)
-            make.right.equalToSuperview().offset(-15)
-            make.height.equalTo(400)
-            make.bottom.equalToSuperview().offset(-15)
-        }
-    }
-}
-
-// MARK: - Private Functions
-
-private extension RunningResultViewController {
-    func configureMap() {
-        self.mapView.delegate = self
-        self.mapView.mapType = .standard
-    }
-    
-    func configureUI() {
+    func configureCommonUI() {
         self.view.backgroundColor = .systemBackground
-        
-        self.configureCommonSection()
-        self.configureDifferentSection()
-    }
-    
-    func configureCommonSection() {
         self.configureScrollView()
         self.configureCloseButton()
         self.configureDateTimeLabel()
@@ -202,6 +166,7 @@ private extension RunningResultViewController {
             
             make.top.equalTo(self.korDateTimeLabel.snp.bottom)
             make.left.equalToSuperview().offset(15)
+            make.right.equalToSuperview().offset(-15)
         }
     }
     
@@ -219,107 +184,5 @@ private extension RunningResultViewController {
             make.left.equalToSuperview().inset(15)
             make.top.equalTo(self.upperSeparator.snp.bottom)
         }
-    }
-    
-    func configureMapViewLocation(from region: Region) {
-        let coordinateLocation = region.center
-        let spanValue = MKCoordinateSpan(latitudeDelta: region.span.0, longitudeDelta: region.span.1)
-        let locationRegion = MKCoordinateRegion(center: coordinateLocation, span: spanValue)
-        self.mapView.setRegion(locationRegion, animated: true)
-    }
-    
-    func showAlert() {
-        let message = "달리기 결과 저장 중 오류가 발생했습니다."
-        
-        let alert = UIAlertController(title: "오류 발생", message: message, preferredStyle: .alert)
-        let cancel = UIAlertAction(title: "취소", style: .default, handler: { _ in
-            self.viewModel?.alertConfirmButtonDidTap()
-        })
-        alert.addAction(cancel)
-        present(alert, animated: false, completion: nil)
-    }
-    
-    func bindViewModel() {
-        guard let viewModel = self.viewModel else { return }
-        let input = RunningResultViewModel.Input(
-            viewDidLoadEvent: Observable<Void>.just(()).asObservable(),
-            closeButtonDidTapEvent: self.closeButton.rx.tap.asObservable()
-        )
-        let output = viewModel.transform(input, disposeBag: self.disposeBag)
-        
-        self.bindMap(with: output)
-        self.bindLabels(with: output)
-        self.bindAlert(with: output)
-    }
-    
-    func bindAlert(with viewModelOutput: RunningResultViewModel.Output) {
-        viewModelOutput.saveFailAlertShouldShow
-            .asDriver(onErrorJustReturn: false)
-            .drive(onNext: { [weak self] _ in
-                self?.showAlert()
-            })
-            .disposed(by: self.disposeBag)
-    }
-    
-    func bindMap(with viewModelOutput: RunningResultViewModel.Output) {
-        viewModelOutput.points
-            .asDriver(onErrorJustReturn: [])
-            .drive(onNext: { [weak self] points in
-                let lineDraw = MKPolyline(coordinates: points, count: points.count)
-                self?.mapView.addOverlay(lineDraw)
-            })
-            .disposed(by: self.disposeBag)
-        
-        viewModelOutput.region
-            .asDriver(onErrorJustReturn: Region())
-            .drive(onNext: { [weak self] region in
-                self?.configureMapViewLocation(from: region)
-            })
-            .disposed(by: self.disposeBag)
-    }
-    
-    func bindLabels(with viewModelOutput: RunningResultViewModel.Output) {
-        viewModelOutput.dateTime
-            .asDriver(onErrorJustReturn: "Error")
-            .drive(self.dateTimeLabel.rx.text)
-            .disposed(by: self.disposeBag)
-        
-        viewModelOutput.dayOfWeekAndTime
-            .asDriver(onErrorJustReturn: "Error")
-            .drive(self.korDateTimeLabel.rx.text)
-            .disposed(by: self.disposeBag)
-        
-        viewModelOutput.mode
-            .asDriver(onErrorJustReturn: "Error")
-            .drive(self.runningModeLabel.rx.text)
-            .disposed(by: self.disposeBag)
-        
-        viewModelOutput.distance
-            .asDriver(onErrorJustReturn: "Error")
-            .drive(self.distanceLabel.rx.text)
-            .disposed(by: self.disposeBag)
-        
-        viewModelOutput.calorie
-            .asDriver(onErrorJustReturn: "Error")
-            .drive(self.calorieLabel.rx.text)
-            .disposed(by: self.disposeBag)
-        
-        viewModelOutput.time
-            .asDriver(onErrorJustReturn: "Error")
-            .drive(self.timeLabel.rx.text)
-            .disposed(by: self.disposeBag)
-    }
-}
-
-extension RunningResultViewController: MKMapViewDelegate {
-    func mapView(_ mapView: MKMapView, rendererFor overlay: MKOverlay) -> MKOverlayRenderer {
-        guard let polyLine = overlay as? MKPolyline else { return MKOverlayRenderer() }
-        
-        let renderer = MKPolylineRenderer(polyline: polyLine)
-        renderer.strokeColor = .mrPurple
-        renderer.lineWidth = 5.0
-        renderer.alpha = 1.0
-        
-        return renderer
     }
 }
